+++ contracts/plugins/assets/abstract/Collateral.sol
@@ -34,7 +34,7 @@ abstract contract Collateral is ICollateral, Asset, Context {
 
     constructor(
         IERC20Metadata erc20_,
-        uint256 maxTradeVolume_,
+        int192 maxTradeVolume_,
         int192 defaultThreshold_,
         uint256 delayUntilDefault_,
         IERC20Metadata referenceERC20_,
