+++ contracts/plugins/mocks/AavePricedFiatCollateralMock.sol
@@ -10,7 +10,7 @@ contract AavePricedFiatCollateralMock is AaveOracleMixin, Collateral {
 
     constructor(
         IERC20Metadata erc20_,
-        uint256 maxTradeVolume_,
+        int192 maxTradeVolume_,
         int192 defaultThreshold_,
         uint256 delayUntilDefault_,
         IComptroller comptroller_,
